requirements.txt
Python>=3.9
ccxt>=4.3.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
redis>=5.0.0
python-dotenv>=1.0.0
ta-lib>=0.4.25
pandas-ta>=0.3.14
asyncio>=3.4.3
aiohttp>=3.8.0
joblib>=1.3.0
requests>=2.31.0
python-telegram-bot>=20.0.0
plotly>=5.15.0
streamlit>=1.28.0
matplotlib>=3.7.0
loguru>=0.7.0
pyyaml>=6.0.0
pytest>=7.4.0
psutil>=5.9.0</parameter>
</write_to_file>